Initial Commit - getball.py
No Raspberry Pi communication with roboRIO. We will implement SSH to execute Robot commands with Java. Possible resource: https://www.chiefdelphi.com/t/executing-files-by-sshing-into-the-roborio/148131/2.

The ball detection works, albeit poorly. I think it is adequate for our use case though and it would be great if it could be tested to see if the ball detection works for the yellow FRC ball when it is put directly in front of the webcam. 

To test, 
1. SSH into the Raspberry Pi
2. Load the script with FileZilla's SCP Function (If you are using Windows) (macOS or Linux has a SCP client built in)
3. Use Pypi to install the opencv-python package.
4. Run script with python3 getball.py and see if console prints True. If it does so satisfactorily, the program is functional.